+++ src/app/components/leaves/leaves.component.ts
@@ -6,7 +6,7 @@ import {
   GroupModel,
   RenderCellEventArgs,
   ScheduleComponent,
-  TimelineMonthService
+  TimelineMonthService, ToolbarActionArgs
 } from '@syncfusion/ej2-angular-schedule';
 import {extend, Internationalization} from '@syncfusion/ej2-base';
 import {LeavesService} from '../../services/leaves/leaves.service';
@@ -76,6 +76,7 @@ export class LeavesComponent implements OnInit {
 
   ngOnInit() {
     this.fetchData();
+
   }
 
   get getPendingApprovalsBadge(): string {
@@ -162,6 +163,7 @@ export class LeavesComponent implements OnInit {
             allowDeleting: false,
             allowEditing: false,
           };
+          // this.schedulerObject.scrollTo('0900');
         }
       } else {
         this.toastComponent.showToast(ToastType.error, 'Error',
@@ -187,10 +189,12 @@ export class LeavesComponent implements OnInit {
         '<div class="content" matTooltip="Leaves left over from last year">Prv. Year</div>';
     }
 
-    if ((args.elementType === 'monthCells' || args.elementType === 'dateHeader') && args.date &&
-      (args.date.getDay() === 0 || args.date.getDay() === 6 || this.checkForHolidayInclude(args.date))) {
+    if ((args.elementType === 'monthCells' || args.elementType === 'dateHeader')) {
       const element = args.element;
-      element.setAttribute('style', 'background-color: ' + leavesColorCombination.holidays);
+      if (args.date && (args.date.getDay() === 0 || args.date.getDay() === 6 || this.checkForHolidayInclude(args.date))) {
+        element.setAttribute('style', `background-color:${leavesColorCombination.holidays}`);
+      }
+
     }
   }
 
@@ -253,4 +257,34 @@ export class LeavesComponent implements OnInit {
   //  Scheduler event handlers and callbacks
   //  ***********************************************************************
 
+  onCreated() {
+    const date: Date = this.schedulerObject.selectedDate;
+    if (date.setHours(0, 0, 0, 0) === new Date().setHours(0, 0, 0, 0)) {
+      this.scrollTo();
+    }
+  }
+
+  onTodayClick() {
+    const date: Date = this.schedulerObject.selectedDate;
+    if (date.setHours(0, 0, 0, 0) === new Date().setHours(0, 0, 0, 0)) {
+      this.scrollTo();
+    }
+  }
+
+  scrollTo() {
+    const today: number = new Date(new Date().setHours(0, 0, 0, 0)).getTime();
+    const todayElement = this.schedulerObject.element.querySelector('[data-date="' + today + '"]') as HTMLElement;
+    const contentWrap = this.schedulerObject.element.querySelector('.e-content-wrap') as HTMLElement;
+    contentWrap.scrollLeft = todayElement.offsetLeft;
+  }
+
+  onActionBegin(args: ToolbarActionArgs) {
+    if (args.requestType === 'toolbarItemRendering') {
+      for (let i = 0; i < args.items.length; i++) {
+        if (args.items[i].text === 'Today') {
+          args.items[i].click = this.onTodayClick.bind(this);
+        }
+      }
+    }
+  }
 }